app/requirements.txt
# FastAPI untuk web API
fastapi==0.115.6

# Celery untuk task queue management
celery==5.4.0

# Tambahan
uvicorn==0.34.0
numpy==1.24.1
imgkit==1.2.3
pillow==11.0.0
python-multipart==0.0.20